Separate actions tests and create an extractor for actions as objects
